fix: refactoring load resolved remotes
