+++ libs/vite-plugin-zephyr/src/lib/internal/mf-vite-etl/load_resolved_remotes.ts
@@ -1,8 +1,7 @@
-import type { FederationRuntimePlugin } from '@module-federation/runtime';
-import acorn from 'acorn';
-import walk from 'acorn-walk';
 import type { ZeResolvedDependency } from 'zephyr-agent';
 import { ze_log } from 'zephyr-agent';
+import { parseRuntimePlugin } from './runtime_plugins_parser';
+import { generateRuntimePlugin } from './runtime_plugin';
 
 export function load_resolved_remotes(
   resolved_remotes: ZeResolvedDependency[],
@@ -11,51 +10,11 @@ export function load_resolved_remotes(
   const startTime = Date.now();
 
   try {
-    // Parse the code into an AST
-    const ast = acorn.parse(code, {
-      ecmaVersion: 'latest',
-      sourceType: 'module',
-      locations: true,
-      ranges: true,
-    });
+    const runtimePluginsExtraction = parseRuntimePlugin(code);
 
-    let pluginsArrayNode: acorn.Node | undefined;
+    if (!runtimePluginsExtraction) return code;
 
-    // Find the plugins array in the runtimeInit call
-    walk.simple(ast, {
-      CallExpression(node: any) {
-        if (node.callee?.name === 'runtimeInit' && node.arguments?.length > 0) {
-          const initArg = node.arguments[0];
-
-          if (initArg.type === 'ObjectExpression') {
-            for (const prop of initArg.properties) {
-              if (
-                prop.key.type === 'Identifier' &&
-                prop.key.name === 'plugins' &&
-                prop.value.type === 'ArrayExpression'
-              ) {
-                pluginsArrayNode = prop.value;
-                break;
-              }
-            }
-          }
-        }
-      },
-    });
-
-    if (
-      !pluginsArrayNode ||
-      !('start' in pluginsArrayNode) ||
-      !('end' in pluginsArrayNode)
-    ) {
-      ze_log('Could not find plugins array in remote entry');
-      return code;
-    }
-
-    // Extract the plugins array
-    const pluginsArrayStart = pluginsArrayNode.start;
-    const pluginsArrayEnd = pluginsArrayNode.end;
-    const pluginsArray = code.slice(pluginsArrayStart, pluginsArrayEnd);
+    const { pluginsArray, startIndex, endIndex } = runtimePluginsExtraction;
 
     // Add Zephyr plugin to the array
     // We need to add a Zephyr plugin to the end of the array
@@ -73,9 +32,7 @@ export function load_resolved_remotes(
 
     // Replace the original array with the updated one
     const updatedCode =
-      code.substring(0, pluginsArrayStart) +
-      updatedPluginsArray +
-      code.substring(pluginsArrayEnd);
+      code.substring(0, startIndex) + updatedPluginsArray + code.substring(endIndex);
 
     const endTime = Date.now();
     ze_log(`load_resolved_remotes took ${endTime - startTime}ms`);
@@ -84,50 +41,4 @@ export function load_resolved_remotes(
     ze_log('Error in load_resolved_remotes:', error);
     return code; // Return original code in case of error
   }
-}
-
-function generateRuntimePlugin(resolved_remotes: ZeResolvedDependency[]): string {
-  const remoteMap = Object.fromEntries(
-    resolved_remotes.map((remote) => [remote.name, remote])
-  );
-  const runtimePlugin: FederationRuntimePlugin = {
-    name: 'zephyr-runtime-remote-resolver',
-    beforeInit: (args) => {
-      const resolvedRemoteMap: Record<string, ZeResolvedDependency> =
-        JSON.parse('__REMOTE_MAP__');
-      const _windows = typeof window !== 'undefined' ? window : globalThis;
-      args.userOptions.remotes.forEach((remote) => {
-        const resolvedRemote = resolvedRemoteMap[remote.name];
-        if (!resolvedRemote) return;
-        const sessionEdgeURL = _windows.sessionStorage.getItem(
-          resolvedRemote.application_uid
-        );
-
-        const urlOverwrite = sessionEdgeURL ?? resolvedRemote.remote_entry_url;
-
-        // @ts-expect-error overwriting entry if needed
-        remote.entry = urlOverwrite;
-      });
-      return args;
-    },
-  };
-  return objectToTemplate(runtimePlugin).replace(
-    '__REMOTE_MAP__',
-    JSON.stringify(remoteMap)
-  );
-}
-
-function objectToTemplate(obj: FederationRuntimePlugin): string {
-  const entries = Object.entries(obj).map(([key, value]) => {
-    if (typeof value === 'function') {
-      return `${key}: ${value.toString()}`;
-    } else if (typeof value === 'object' && value !== null) {
-      return `${key}: ${objectToTemplate(value)}`;
-    } else if (typeof value === 'string') {
-      return `${key}: "${value}"`;
-    } else {
-      return `${key}: ${value}`;
-    }
-  });
-  return `{ ${entries.join(', ')} }`;
 }
